clean up shared test code structure
